+++ tests/ComfortPay/GatewayTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Omnipay\TatraPay;
+namespace Omnipay\ComfortPay;
 
 use Omnipay\Tests\GatewayTestCase;
 
@@ -18,34 +18,6 @@ class GatewayTest extends GatewayTestCase
         $this->gateway->setMid(1111);
     }
 
-    public function testPurchaseSignWithDes()
-    {
-        if (version_compare(PHP_VERSION, '7.4.0', '>=')) {
-            $this->markTestSkipped('Skipped for php 7.4');
-        }
-
-        $this->gateway->setSharedSecret('11111111');
-
-        $request = $this->gateway->purchase(array(
-            'amount' => '10.00',
-            'currency' => 'EUR',
-            'vs' => 123456,
-            'rurl' => 'http://return.sk',
-        ));
-        
-        $this->assertInstanceOf('Omnipay\TatraPay\Message\PurchaseRequest', $request);
-        $this->assertSame('10.00', $request->getAmount());
-
-        $response = $request->send();
-
-        $this->assertTrue($response->isRedirect());
-
-        $this->assertEquals(
-            'https://moja.tatrabanka.sk/cgi-bin/e-commerce/start/e-commerce.jsp?PT=TatraPay&MID=1111&CURR=978&VS=123456&AMT=10.00&RURL=http%3A%2F%2Freturn.sk&SIGN=C758BF5F3A60FC3B',
-            $response->getRedirectUrl()
-        );
-    }
-
     public function testPurchaseSignWithAes()
     {
         if (version_compare(PHP_VERSION, '7.4.0', '>=')) {
@@ -59,9 +31,11 @@ class GatewayTest extends GatewayTestCase
             'currency' => 'EUR',
             'vs' => 123456,
             'rurl' => 'http://return.sk',
+            'ipc' => 'a',
+            'name' => 'test',
         ));
         
-        $this->assertInstanceOf('Omnipay\TatraPay\Message\PurchaseRequest', $request);
+        $this->assertInstanceOf('Omnipay\ComfortPay\Message\PurchaseRequest', $request);
         $this->assertSame('10.00', $request->getAmount());
 
         $response = $request->send();
@@ -69,7 +43,7 @@ class GatewayTest extends GatewayTestCase
         $this->assertTrue($response->isRedirect());
 
         $this->assertEquals(
-            'https://moja.tatrabanka.sk/cgi-bin/e-commerce/start/e-commerce.jsp?PT=TatraPay&MID=1111&CURR=978&VS=123456&AMT=10.00&RURL=http%3A%2F%2Freturn.sk&SIGN=A50341DBBE65CB0A0E67F487C463DAF9',
+            'https://moja.tatrabanka.sk/cgi-bin/e-commerce/start/e-commerce.jsp?PT=CardPay&MID=1111&CURR=978&VS=123456&AMT=10.00&AREDIR=1&RURL=http%3A%2F%2Freturn.sk&IPC=a&NAME=test&TPAY=Y&TEM=1&SIGN=CA6D5581146815F7BCDD3072D39B481A',
             $response->getRedirectUrl()
         );
     }
@@ -83,12 +57,15 @@ class GatewayTest extends GatewayTestCase
             'currency' => 'EUR',
             'vs' => 123456,
             'rurl' => 'http://return.sk',
+            'ipc' => 'a',
+            'name' => 'test',
+            'terminalId' => 1,
         ));
 
         $request->setTimestamp('01022021214520');
 
         
-        $this->assertInstanceOf('Omnipay\TatraPay\Message\PurchaseRequest', $request);
+        $this->assertInstanceOf('Omnipay\ComfortPay\Message\PurchaseRequest', $request);
         $this->assertSame('10.00', $request->getAmount());
 
         $response = $request->send();
@@ -96,7 +73,7 @@ class GatewayTest extends GatewayTestCase
         $this->assertTrue($response->isRedirect());
 
         $this->assertEquals(
-            'https://moja.tatrabanka.sk/cgi-bin/e-commerce/start/tatrapay?PT=TatraPay&MID=1111&CURR=978&VS=123456&AMT=10.00&RURL=http%3A%2F%2Freturn.sk&TIMESTAMP=01022021214520&HMAC=949f80a0b9a5626bc4a742de1835d7f307e8690a0eeb43473c5f79405023f43a',
+            'https://moja.tatrabanka.sk/cgi-bin/e-commerce/start/cardpay?PT=CardPay&MID=1111&CURR=978&VS=123456&AMT=10.00&AREDIR=1&RURL=http%3A%2F%2Freturn.sk&IPC=a&NAME=test&TPAY=Y&TEM=1&TIMESTAMP=01022021214520&HMAC=ffa9b92506ae87bff4bdbd7ef95d4877a37af79a1dc839585add56068083dc03',
             $response->getRedirectUrl()
         );
     }